Refactor SearchOverlay to use Context for state management (#58768)
